+++ src/hooks/GJUserCell.cpp
@@ -1,4 +1,4 @@
-#include <Geode/binding/GameManager.hpp>
+#include "../AnimatedIcons.hpp"
 #include <Geode/binding/GJSpiderSprite.hpp>
 #include <Geode/binding/SimplePlayer.hpp>
 #include <Geode/modify/GJUserCell.hpp>
@@ -7,12 +7,7 @@ using namespace geode::prelude;
 
 class $modify(APIUserCell, GJUserCell) {
     struct Fields {
-        ~Fields() {
-            auto gameManager = GameManager::get();
-            gameManager->setUserObject("original-scale"_spr, nullptr);
-            gameManager->setUserObject("selected-child"_spr, nullptr);
-            gameManager->setUserObject("touch-children"_spr, nullptr);
-        }
+        ~Fields() { AnimatedIcons::releaseTouchChildren(); }
     };
 
     void loadFromScore(GJUserScore* score) {
@@ -30,7 +25,7 @@ class $modify(APIUserCell, GJUserCell) {
                 player->m_spiderSprite->setUserObject("reset"_spr, CCBool::create(true));
                 touchChildren->addObject(player->m_spiderSprite);
             }
-            if (touchChildren->count() > 0) GameManager::get()->setUserObject("touch-children"_spr, touchChildren);
+            AnimatedIcons::addTouchChildren(touchChildren);
         }
 
         m_fields.self();