+++ src/hooks/ItemInfoPopup.cpp
@@ -1,4 +1,4 @@
-#include <Geode/binding/GameManager.hpp>
+#include "../AnimatedIcons.hpp"
 #include <Geode/binding/GJSpiderSprite.hpp>
 #include <Geode/binding/SimplePlayer.hpp>
 #include <Geode/modify/ItemInfoPopup.hpp>
@@ -7,12 +7,7 @@ using namespace geode::prelude;
 
 class $modify(APIItemInfoPopup, ItemInfoPopup) {
     struct Fields {
-        ~Fields() {
-            auto gameManager = GameManager::get();
-            gameManager->setUserObject("original-scale"_spr, nullptr);
-            gameManager->setUserObject("selected-child"_spr, nullptr);
-            gameManager->setUserObject("touch-children"_spr, nullptr);
-        }
+        ~Fields() { AnimatedIcons::releaseTouchChildren(); }
     };
 
     bool init(int id, UnlockType type) {
@@ -24,7 +19,7 @@ class $modify(APIItemInfoPopup, ItemInfoPopup) {
             auto touchChildren = CCArray::create();
             if (type == UnlockType::Robot && player->m_robotSprite) touchChildren->addObject(player->m_robotSprite);
             if (type == UnlockType::Spider && player->m_spiderSprite) touchChildren->addObject(player->m_spiderSprite);
-            if (touchChildren->count() > 0) GameManager::get()->setUserObject("touch-children"_spr, touchChildren);
+            AnimatedIcons::addTouchChildren(touchChildren);
         }
 
         m_fields.self();