+++ changelog.md
@@ -1,5 +1,6 @@
 # Animated Icons Changelog
 ## v1.0.1 (2025-02-10)
+- Added animation to the second player in the icon kit (If the mod "Separate Dual Icons" by Weebify is enabled)
 - Fixed a bug where clicking on a robot or spider would make it disappear
 
 ## v1.0.0 (2025-02-09)